Add files via upload
This is a photo of what was captured on the oscilloscope during a single exposure/capture.  The yellow signal is the LED (HIGH to LOW) and the blue is the signal of the Input2 on the Baslar. It's clear after multiple runs that the exposure window will always within the LED exposure using a software trigger.